refactor: rename to CommonNodeClass
