+++ pkg/apis/v1beta1/commonnodeclass.go
@@ -1,0 +1,45 @@
+/*
+Portions Copyright (c) Microsoft Corporation.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1beta1
+
+import (
+	"github.com/awslabs/operatorpkg/status"
+	corev1 "k8s.io/api/core/v1"
+)
+
+// NodeImage contains resolved image selector values utilized for node launch
+type NodeImage struct {
+	// id is the ID of the image. Examples:
+	// - CIG: /CommunityGalleries/AKSUbuntu-38d80f77-467a-481f-a8d4-09b6d4220bd2/images/2204gen2containerd/versions/2022.10.03
+	// - SIG: /subscriptions/10945678-1234-1234-1234-123456789012/resourceGroups/AKS-Ubuntu/providers/Microsoft.Compute/galleries/AKSUbuntu/images/2204gen2containerd/versions/2022.10.03
+	// +required
+	//nolint:kubeapilinter // requiredfields: validation is intentionally not enforced for this field
+	ID string `json:"id"`
+	// requirements of the image to be utilized on an instance type
+	// +required
+	//nolint:kubeapilinter // requiredfields: omitempty is intentionally omitted for this field
+	Requirements []corev1.NodeSelectorRequirement `json:"requirements"`
+}
+
+type CommonNodeClass interface {
+	StatusConditions() status.ConditionSet
+	GetKubernetesVersion() (string, error)
+	GetImages() ([]NodeImage, error)
+	Hash() string
+	GetName() string
+	GetAnnotations() map[string]string
+}
